MainActivity not to finish while assistant app in config
Assistant app setting is called from config, which the user may come
back to config and then MainActivity.

In current implementation, MainActivity already has finished when user
has come back to config.

In this commit MainActivity will not finish in that situation. Going to
assist app setting but not coming back to config is also a possible way.
At least in my testing it works properly also in that situation.