Remove duplicated icons from fastlane icons
These 2 files are not needed because it is identical to English version,
and F-Droid fallbacks to English version.

The document says that en-US is the F-Droid fallback language, which
implies some kind of fallback works.
https://f-droid.org/ja/docs/All_About_Descriptions_Graphics_and_Screenshots/

This is suggested by a pull request the below:
https://github.com/nhirokinet/bluelineconsole/pull/280
but the author of the pull request looks like not responding, and the
pull request could not be merged because the commits remove files which
must not be removed. For this reason, this commit is made separately.