Update MainActivity.java
header and footer texts are now empty depending on the related shared preference (by changes to the onCreate function). onResume function now sets the header and footer texts as the configuration might be changed. it also calls resultCandidateListAdapter.notifyDataSetChanged() , since the configuration might be changed related to the visibility of the app icons.